Created print_conversion function and tested it.
This function will take in the lines of Grin commands as tokens and check if its first token is PRINT. If it is it will append the value that wants to be printed to a list. It returns a list of print values for testing, but also displays the print in the console.