Updated let_conversion and take_user_grin_input
For let_conversion, when you do a grin command such that "LET X Y", it would only store X in the list of values. I made it so it also stores Y in the list of values as well, both being 0. As for take_user_grin_input, changed so it only stops taking inputs when given a '.', not 'END'