Finished jump_line function and tested a majority of it. Created check_condition function and partially tested it.
For the jump_line function, I made it jump accordingly when it encounters an Identifier. If its an invalid identifier, where label doesn't exist or float or 0,it will return an error. check_condition will check the conditional of goto if given one. Only tested with identifier and literal int.